Refactored omas_machine to remove ringinclusion - Moved functions needed by mappings to seprate file   - Moved the machine mapping decorator to a utilities file   - Moved MDS+ related routines into their own file - Added extra argument to pass `update_mapping` as an argument - Removed `compile` statement as it was causing trouble for the debugger
