+++ omas/machine_mappings/_common.py
@@ -3,6 +3,7 @@ from omas import *
 from omas.omas_utils import printd
 import os
 import glob
+from omas.omas_setup import omas_dir
 
 __support_files_cache__ = {}
