+++ omas/machine_mappings/d3d.py
@@ -6,6 +6,10 @@ from numpy.lib.function_base import iterable
 from omas import *
 from omas.omas_utils import printd, printe, unumpy
 from omas.machine_mappings._common import *
+from uncertainties import unumpy
+from omas.utilities.machine_mapping_decorator import machine_mapping_function
+from omas.utilities.omas_mds import mdsvalue
+from omas.omas_core import ODS
 
 __all__ = []
 __regression_arguments__ = {'__all__': __all__}
